Extract constraining the video to its own class
